add user authentication with JWT, update profiles.postman_collection.json

+++ src/main/java/com/telros/profiles/controller/ProfileController.java
@@ -2,13 +2,16 @@ package com.telros.profiles.controller;
 
 import com.telros.profiles.DTO.ProfileContactsDTO;
 import com.telros.profiles.DTO.ProfileDTO;
-import com.telros.profiles.request.AddProfileRequest;
+import com.telros.profiles.model.Profile;
 import com.telros.profiles.request.EditProfileContactsRequest;
 import com.telros.profiles.request.EditProfileRequest;
 import com.telros.profiles.service.ProfileService;
 import jakarta.validation.Valid;
 import java.util.List;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.HttpClientErrorException;
 
 @RestController
 public class ProfileController {
@@ -38,25 +41,38 @@ public class ProfileController {
     return profileService.getProfileContacts(id);
   }
 
-  @PostMapping(value = "/profiles")
-  public Long addProfile(@RequestBody @Valid AddProfileRequest request) {
-    return profileService.addProfile(request);
-  }
-
   @PutMapping(value = "/profiles/{id}")
   public Long editProfile(@PathVariable Long id,
-                          @RequestBody @Valid EditProfileRequest request) {
-    return profileService.editProfile(id, request);
+                          @RequestBody @Valid EditProfileRequest request,
+                          @AuthenticationPrincipal Profile profile) {
+    if (profile.getId().equals(id)) {
+      return profileService.editProfile(id, request);
+    } else {
+      throw new HttpClientErrorException(HttpStatus.FORBIDDEN,
+          "Вы не можете редактировать чужой профиль");
+    }
   }
 
   @PutMapping(value = "profiles/contacts/{id}")
   public Long editProfileContacts(@PathVariable Long id,
-                                  @RequestBody @Valid EditProfileContactsRequest request) {
-    return profileService.editProfileContacts(id, request);
+                                  @RequestBody @Valid EditProfileContactsRequest request,
+                                  @AuthenticationPrincipal Profile profile) {
+    if (profile.getId().equals(id)) {
+      return profileService.editProfileContacts(id, request);
+    } else {
+      throw new HttpClientErrorException(HttpStatus.FORBIDDEN,
+          "Вы не можете редактировать чужой профиль");
+    }
   }
 
   @DeleteMapping(value = "profiles/{id}")
-  public Long deleteProfile(@PathVariable Long id) {
+  public Long deleteProfile(@PathVariable Long id,
+                            @AuthenticationPrincipal Profile profile) {
+    if (profile.getId().equals(id)) {
     return profileService.deleteProfile(id);
+    } else {
+      throw new HttpClientErrorException(HttpStatus.FORBIDDEN,
+          "Вы не можете удалить чужой профиль");
+    }
   }
 }
